+++ src/Table/LaravelExcelException.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace ProtoneMedia\Splade\Table;
+
+use Exception;
+
+class LaravelExcelException extends Exception
+{
+}
